mod_webmail: Fix RFC822 date parsing.
The existing date parsing handled most cases
of dates in email headers, but not all of them;
in particular, cases where just the TZ abbreviation,
as opposed to an offset, was present were not
properly parsed. This case is now handled.

Additionally, improper ordering made it possible
to set the file activity of interest to 0 (since
it had not yet been initialized), which would
erroneously trigger an activity callback whenever
the sysop console received input (STDIN). This is
also fixed.