net_imap: Eliminate nonexistent SEARCH capability.
There is no such thing as the SEARCH capability;
SEARCH is part of the base IMAP4rev1 specification.
(There are SEARCH extensions that build on SEARCH,
but there is no such capability for base SEARCH.)

Therefore, it's wrong to require such a capability
to be supported by remote servers to pass through
SEARCH commands.