mod_webmail: Fix SEARCH/SORT command construction.
This fixes the SEARCH and SORT commands (in both
cases, the search key part) to construct proper
RFC 3501 SEARCH commands using RECENT and UNSEEN,
rather than using KEYWORD and UNKEYWORD, since this
is not to spec and not supported by any IMAP server.

This was done initially as sort of a kludge as it
had seemed that libetpan did not support UNSEEN
or RECENT. In practice, it does, and although code
is now present to do it the proper way for that library,
it does not seem to work properly due to the search key
types being converted into the wrong string when the
command is sent. This seems almost like an ABI issue,
given the off by 1/off by 2 type of issue, but there
is no actual evidence to back this up.

In the meantime, we construct the command manually
to ensure that we properly construct a valid (and
the intended) command. If the libetpan bug is figured
out in the future, the alternate code can be removed.

Additionally, since some code was added to net_imap
purely to support the kludge of using KEYWORD and
UNKEYWORD, and beacuse this is not valid IMAP syntax
and is no longer done in mod_webmail, that code is
now removed.