net_imap: Fix stale remote LIST attributes.
Previously, LIST responses for remote IMAP servers
(client proxies) were cached on disk, so that
LIST responses could be provided immediately without
having to establish remote connections at all.

However, this is not correct, because not all mailbox
attributes are fixed. In particular, the Marked and
Unmarked attributes could change at any time, and
the only way to get the correct attributes is to
ask for them in realtime. Therefore, we now do
a live LIST command for every remote server when
requested.

Thankfully, due to the new parallel framework,
this is much more performant than it would have
been in the past. Additionally, this also
greatly simplifies much of the LIST and LIST-STATUS
handling, as it allows the LIST-STATUS inquiries
to be satisfied in a much more elegant manner
than the two-phase approach that was used
previously.