mod_webmail: Read entire multiline IDLE responses.
We can currently read a single line of untagged
data received during IDLE, but not more than that.
Calling mailimap_read_line repeatedly works great
to retrieve successive lines of IDLE data, but
it is blocking so it must not be called unless
we're sure that there is more data to read, so
this fix is a little creative on that front.