mod_webmail: Allow selecting all messages in a mailbox.
Previously, it was only possible for the frontend to specify
messages that were the target of an operation (e.g. MOVE, COPY,
STORE, etc.) by explicitly enumerating all the relevant
UIDs in a provided array. However, there are certain cases
where clients may want to do something to all messages in the
entire mailbox, so we now support 1:* operations.

Other fixes enclosed:
* Don't idle if we're already idling
* Issue NOOP prior to IDLE on empty mailbox to prevent
  lack of parser synchronization in libetpan