net_imap: Fix IDLE untagged EXISTS for non-INBOX.
Untagged EXISTS responses during IDLE have previously
only worked for the INBOX. This was due to a hardcoded
assumption that the INBOX (root maildir for a mailbox)
would be the directory of interest. This is not necessarily
true for APPEND actions, so we now use the actual directory
and untagged EXISTS for IDLE now works properly for all mailboxes.