net_imap: Inject NOMODSEQ response for remote mailboxes if needed.
RFC 4551 says we must send a NOMODSEQ response during SELECT
or EXAMINE if a selected mailbox does not support persistent
modsequences. For remote mailbox proxying, if the remote server
does not support MODSEQ (but the client does), then we need to
dynamically inject this response into the SELECT/EXAMINE reply.