net_imap: Partially extend NOTIFY support to remote servers.
This (partially) adds support for NOTIFY for remotely proxied
mailboxes, not by actually doing a remote NOTIFY, but by
running the IDLE command in the background on the remote
server. We also now actively monitor all connections during
an IDLE using poll, so we can be immediately alerted about
activity on any connection, and we can then translate that
into a NOTIFY-compliant response and pass that on to the
client. This is transparent, but also limited as currently
this functionality only works for the INBOX on each remote
connection. This is a simplification made as IDLE only allows
monitoring one mailbox per connection. This implementation
thus is reasonably good enough and adds no additional overhead.

Other fixes/improvements (not related to remote mailboxes):

* Actually do a traversal for NOTIFY STATUS responses.
* Send untagged RECENT response for new messages with IDLE.