net_imap: Manage and prevent remote disconnections during IDLE.
If a remote IMAP server disconnects our connection while we're
idling, we need to disconnect the client since it's dead, or
poll will keep getting triggered by it. This is most common
with IMAP servers that do not comply with RFC 3501, i.e.
they disconnect clients with fewer than 30 minutes of inactivity
(even when idling!)

Additionally, as a proactive measure to account for servers
that disconnect clients prematurely, there is now builtin
support to do "keepalives" on IMAP client connections during
IDLE (which is actually necessary no matter what, due to
the 30 minute limit, but is also useful in this case for
shorter durations as well).