net_imap: Truly parallelize remote LIST command.
The LIST command for remote mailboxes was set up
using the parallel framework, but the client
was set up synchronously before a job was created
for it, which was rather silly as TCP and IMAP setup
do not take a trivial amount of time. This is now
moved to inside the job itself, and the LIST response
for client proxies no longer has serial bottlenecks.