net_imap: Standardize IDLE for local/remote folders.
Previously, IDLE for remote mailboxes used the
passthru approach that many other commands did,
where net_imap merely proxies the command and
relays the responses back, with no intelligence
added. This was previously perfectly suitable,
but is now not desirable because of the added complexity
of IDLE logic to properly handle NOTIFY, including
doing an IDLE keepalive on all the remote connections.

This eliminates passthru of IDLE since now, IDLE is
automatically invoked on all remote clients that
support it anyways. Therefore, the only thing that
needs to be done specially for remote idling is make
sure that we are, in fact, idling on that mailbox,
and then literally relay any responses received,
rather than doing the usual NOTIFY conversion processing
that's done for remote IDLE activity otherwise.

As a result, NOTIFY is now transparently supported
if a client is idling on a remote mailbox, not just
if idling on a local mailbox.

Additionally, previously the IDLE handler was not
actually checking that a remote mailbox fell within
the NOTIFY watchlist. This is now checked.