net_imap: Fix unresolved symbols for RTLD_LAZY.
net_imap was declining to load even with RTLD_LAZY
due to a function in another module being used as
a variable, rather than as a normal variable, i.e.
weak unresolved. Because we could not lazy load it
to read its dependencies, we failed to add its
dependencies to the preload list, and thus when
net_imap failed to load for real during the loading
process, it was not apparent that there were missing
dependencies that needed to be loaded first.

To fix this, because unresolved variables do not work
with RTLD_LAZY, an in-module wrapper function around
uidsort is added to work around that.

Fixes: #16