net_imap: Optimize remote SIZE recalculations.
This adds an additional optimization to compute
the SIZE of a remote mailbox, if requested (when
the remote server does not support SIZE). Currently,
we only returned the cached value if the mailbox has
not changed at all. Now, if the mailbox has only
grown larger (and no messages have been expunged),
we only calculate the incremental size increase and
add it to the cached size from last time, which
can often result in significant time savings.

This does assume that UIDNEXT increases by 1 for
every message added to a mailbox, but this is
standard behavior and unlikely to pose an issue.