net_imap: Fix improper loop continue.
continue is inside a double loop here,
and it was continuing the inner loop,
rather than the outer loop as intended.

This caused remote clients that received
IDLE activity to loop more than they
should have here and then attempting
to terminate the IDLE would fail due
to lack of synchronization.