net_imap: Fix NOTIFY permission checks.
Mailbox permissions were not being properly
verified for NOTIFY notifications. While
generate_status does do a mailbox traversal,
which does an ACL check, this is only done
once for each notification as an optimization,
and thereafter the STATUS response is reused
for any remaining sessions. Therefore, the ACL
for each session still needs to be explicitly
checked. Rather than regenerating the STATUS
response every time just to do the ACL check
(which would also work), we incorporate this
as a separate check, which should have less
overhead.