net_imap: Add transparent cross-server COPY/MOVE.
This adds support for COPY and MOVE across IMAP
server boundaries, i.e. between the local server
and a remote server or two different remote
servers. This supplements the client proxy
functionality which allows for presenting remote
mailboxes to the user as if they were local.
However, it has hitherto been impossible to COPY
or MOVE messages where the source and destination
mailboxes were not either both local or both on
the same remote server. Now, FETCH and APPEND are
used transparently to make this possible, which
allows for messages to be easily moved between
servers, much as a user can do in a traditional
offline mail client.