pty.c: Allow output to be interrupted/discarded.
The emulated speed mechanism use slow_write
in PTY to carefully time out the write, but
since this is done all at once, if there is
any input pending for the BBS while the slowed
write is being done, the input will not be
passed to the slave PTY until the previous
output has finished writing.

This allows slow output to be suspended and
resumed if input is pending from the node.
The PTY thread will relay the input, and then
resume writing the previous output. However,
the user is now able to interrupt this output
with ^C, which will simply abort any pending
output, useful for discarding a large output
buffer at a slow output speed. ^Z is also
mapped to this same function, since the
PAUSE/BREAK key seems to map to that.