net_imap: Dynamically calculate most efficient IDLE poll interval.
There are several different "events" or "timeouts" competing
for attention during an IDLE, namely remote IDLE expirations,
the local IDLE expiration, and the ping interval for inactivity.
Instead of trying to use a sufficiently short fixed interval
that accomodates all of these, which inevitably results in an
inappropriately short interval much of the time coupled with
large uncertainity, dynamically calculate when the next event
of interest will occur and use that for the poll interval.
This increases the length of the average poll, thereby also
reducing the amount of spurious debug logging for IDLE polls.