net_imap: Fix overly broad LIST responses.
The LIST response was including mailboxes in
the personal namespace, even for commands like:

a LIST "" "Shared Folders*"

This is because the namespace was previously
used to control whether the Other and Shared
namespaces would be traversed, but the personal
namespace was always traversed and no check
appropriately excluded these non-matching mailboxes.
This is a special case because "Other Users"
and "Shared Folders" are skipped if they appear
at the beginning of the pattern, so the comparison
was just matching on "*", which, of course, would
match anything and everything.

We now also keep track if a namespace is explicitly
associated with the pattern, and if this does not
match the namespace being traversed, it cannot be
a match, and we skip it.

LIST respons generation still has some correctness flaws,
which are noted in a comment for now.