menu, node: Optimize multispace writing.
Formatting is heavily used throughout the
BBS terminal system, particularly for menus.
Because printf-style formatting is used for
alignment, this results in large amounts of
space padding, which is inefficient to send
on the wire. For ANSI-supporting nodes, we
now use ANSI escape sequences to "skip N
columns" rather than sending N spaces explicitly,
when it is more efficient to do so. This results
in a significant performance speedup on low
speed connections (e.g. 300, 1200 bps).