system.c: Add resource limits.
It was originally conceived that cgroups
would be needed to limit resource consumption
used inside containers. However, we can
more or less achieve the same goal by using
rlimits inside the new namespace. This helps
limit the effect of launched programs on
the rest of the BBS (and the system itself).