net_imap: Look for .imapremote in user's home directory.
Now that users can easily edit files in their home
directory, even from a simple BBS terminal session,
look for .imapremote in a user's home directory,
rather than in the user's root maildir. This change is
made because users' maildirs are not directly accessible
to them, but this is a file users will need to create
and edit themselves.

This change is not backwards compatible for users.

This does highlight that mod_oauth.conf cannot be edited
by users and contains any OAuth mappings needed for OAuth
authentication. Ideally, users would also be able to
define OAuth authentication mappings in their home directories,
and mod_oauth would use this configuration instead.

This would also resolve possible security risks when a user ID
is not specified for a mapping, allowing anyone to use the OAuth
credentials of another user. Future work is needed to allow
mod_oauth to look for its configuration in multiple places.
Additionally, the module currently loads this information when
the module is loaded, so an inotify callback would be needed to
get alerted about updates to the file. Alternatively, as with
.imapremote, the mapping need not be stored in memory, but
could be accessed on demand when needed, directly from disk.
The downside of this is tokens would not be reused, since we
need to appropriately keep track of expiration.