mod_oauth: Also use configs in home directories.
This extends mod_oauth to also look for token configs
in the user's home directory, in .oauth.conf, as well
as the existing global mod_oauth.conf config file.
Because OAuth tokens are currently used for IMAP
and SMTP proxying (the former of which is now
user configurable), it makes sense for users to be able
to configure their OAuth credential mappings themselves.

To support this, tokens are now uniquely identified by
their name + userid combo, rather than just the name.
If a token is not found initially, we try loading .oauth.conf
from the user's home directory. We also do this if we
find a mapping but fail to get a token from it, just in
case the user has updated .oauth.conf with new values,
since otherwise the user cannot force changes to take effect.

This improvement mitigates most of the original security
concerns that existed when this module was first written,
since cross-user token usage is now adequately guarded against.