net_imap: Make NOTIFY comparisons case-insensitive.
Currently, "INBOX" is hardcoded to be the mailbox
of interest on backgrounded remote connections,
but some servers spell it "Inbox" instead. Even
though mailboxes on the local IMAP server are not
case insensitive, we use case-insensitive comparisons
for this purpose for compatibility.

This also ensures the transfer subsystem is initialized
even if FTP is not being used, since it is needed for
loading the .imapremote file (as it resides in the user's
home directory).