mod_spamassassin: Add SpamAssassin filtering support.
This adds integration for SpamAssassin to run on
incoming emails. Rather than using a milter as
sendmail and postfix allow, this currently only
supports spawning the spamassassin program for
each message. Future work may refine this.

To do this, the logic for prepending headers to
a received message is abstracted a bit and exposed
to other modules, so this filtering logic can be
isolated outside of the main SMTP module itself.
This should allow future work to be less disruptive.

Logic to add the Received header, Precedence header,
and SPF header have thus been moved into separate
modules that must be loaded to modify the message.