startup.c: Prevent startup callback dependency deadlock.
Startup callbacks are commonly used to finish
initialization or some other task that has been
delayed. In particular, since most TCP sockets
are registered during startup, the TCP multilistener
thread is not notified about each of these since
it is more efficient to notify it just once upon
startup, rather than per each registered TCP socket.
A startup callback is used to do this. However,
other startup callbacks may implicitly depend on this;
if door_irc.conf is configured to connect to localhost,
for example, and the startup callback for the first
TCP listener was registered afterwards, then door_irc
would be unable to connect to localhost since the TCP
multilistener thread isn't yet listening on these sockets,
since that startup callback is later in the callback list.

To fix this, startup callbacks now have priorities.
Most callbacks can use the default priority because
are neither dependent on anything else nor are depended
on by anything else. Certain callbacks, such as in the
example above, may have relationships, and these have
an explicit priority assigned to ensure that callbacks
are run in the correct order.