mod_slack: Add Slack/IRC relay module.
This adds a Slack/IRC relay module that is similar
to mod_discord in functionality, except for Slack.
Additionally, this introduces the concept of a
"personal relay", a relay dedicated to an individual
user for 1:many relay purposes, as opposed to the
existing relays which are many:many relays. This
is applicable to any type of relay, but currently
only present in the Slack module.