net_smtp: Properly handle multiline greetings.
This fixes handling of sending emails to MTAs
that may have a multiline server greeting.
The protocol allows this, and this is commonly
used for tarpitting.

Previously, we only waited for a single 220
response without checking if it was the last one,
and if it was not, we would send the EHLO immediately,
which is a protocol violation since we need to wait
for the entire server greeting. Spam filters like
postscreen flag this as PREGREET and reject mail
at the RCPT TO stage with a Protocol violation error.
Now, we properly wait in a loop, also with a slightly
increased wait time to accomodate tarpitting.