module.c: Fix SEGV during shutdown unload of dependencies.
Attempts to access invalid memory could occur during
shutdown while trying to recursively unload a module's
dependencies. This can happen if module A is required by
modules B and C, and module B is also required by module
C. Unloading A might try to unload B first, which would
unload C before we attempt to unload C due to it depending
on A (rather than B). As a result, C has already been
freed and is no longer valid memory.

To avoid this, we now copy the list of modules to unload
with names and check if the module has already been unloaded
before attempting to unload it. If so, we skip the duplicate
unload, which is not needed and would cause a segfault.