node: Prevent deadlocks due to blocked writes.
Because the BBS mostly uses a relatively simple
threading model (one thread per node, with
blocking I/O), write can block and previously
this has not been handled well. In particular,
a common deadlock observed is that if a node
is blocked on I/O, since the node locked is held,
any attempts to print the node list will deadlock
the entire BBS, since the node traversal locked
each node as it traversed. This is removed as
the per-node locking was not really necessary.

The more comprehensive fix here takes further
steps to ensure that the system as a whole
does not deadlock, even if individual node threads
block due to blocked writes (which is fine,
but should not be allowed to disrupt other
threads).

All node I/O to file descriptors now uses
bbs_node_fd_ functions, which properly serialize
writes. dprintf is no longer used to write to
nodes. If the node thread is doing the write,
the node is locked as usual since this thread
is allowed to block insofar as it is no disruptive
to the rest of the system. The more interesting
case is when a thread that is not handling a node
tries to write to it. This is done, for example,
in the IMAP and IRC servers when the action of
one client causes messages to be sent to other
clients. However, when writing, if that node is
already blocked, then the other thread would also
block, and this is very bad. To mitigate this, rather
than locking the node outright, we now use trylock
and abort if we cannot acquire the lock. Additionally,
non-blocking writes are used in this case to avoid
blocking indefinitely.

Because locking is now handled here, this also simplifies
application usage, since the IMAP and IRC modules no longer
need to provide their own locking to serialize writes to
node file descriptors, and as I/O is more concentrated
to a few core functions, there should be fewer places
that things can go wrong.