startup.c: Simplify startup callback registration.
In many cases, when modules are loading or subsystems
are initializing, there are callbacks that should be
run either immediately or once the BBS is fully started
if it is still starting, either because of dependencies
on other components or to delay doing work that is not
necessary during startup. A common wrapper function is
now added to simplify these kinds of calls.