mod_webmail: Prevent infinite loop if IDLE out of sync.
IDLE can occasionally get out of sync (due to probably
another bug in handling this). However, if we are, currently
the return value is not checked so we may repeatedly try to
start and stop IDLE in a loop, failing every time. Now,
if we ever fail to start IDLE, we will simply abort; if
this happens, future operations are likely to messed up
anyways due to lack of synchronization, so we may as well
abort the session then.