mod_webmail: Fix display of UTF-8 characters.
Previously, the frontend was displaying ??? symbols
instead of UTF-8 characters, wherever they appeared.
The fundamental issue here is that unsigned char
should be used to allow UTF-8 values to be captured,
since these use values 128 through 255. When signed
char is used (as is typically used by default), these
just end up setting the high bit in 2s complement
format, such that the values are negative and this
becomes an invalid string, which also previously
caused JSON encoding to fail.

The proper solution would be to explicitly use
the unsigned char type everywhere needed in this
module. For now, since signed char is not needed
at all in mod_webmail (or really anywhere, in the
BBS, since char is only used for strings and
characters, not bytes values generally), all uses
of char are forced to be unsigned through a gcc
compiler option, which fixes this issue in an
easy way for now.