socket.c: Prevent write from blocking forever.
Improving the work done in commit
1f68ac22a25a4af95149d9514e775d6b6ee17cc9,
if we're trying to do a full write, only
wait up to 60 seconds for a file descriptor
to become writable. If this fails, consider
this a permanent failure. This will fix some
deadlock issues, but is still fine to do since
we are not obligated to wait forever to write.