mod_events: Allow IPs/CIDRs to be whitelisted.
Allow whitelisting of CIDR ranges or individual
IP addresses, to prevent trusted IP addresses
from being blocked by the autoblock functionality.
For example, a trusted server hosting a modem bank
might result in a large number of protocol failures
due to data corruption, but the server should not
be blocked, as it is merely acting as a proxy.
Likewise, a sysop will want to prevent him/herself
from accidentally getting locked out of the server.