mod_webmail: Fix marking entire mailbox as read/unread.
When the frontend marks an entire mailbox as seen or unseen,
it doesn't have enough information to know what the new
unseen count for that mailbox should be, since it doesn't
know what the flags are of messages not on the current page.

Therefore, we now do an implicit STATUS on 1:* operations,
and send a STATUS update to the frontend, so that it can
update the counts in the folder pane. We also explicitly
refresh the current page, although the frontend could
also just mark everything on the current page as seen/unseen.