net_imap: Fix duplicate client cleanup.
If we exceed the number of allowed remote clients,
we need to remove an existing client to add a new
one. However, this was not being done in a safe way,
resulting in a client being destroyed while it was
being used, since this code was written prior to
parallel client usage being possible. Now, clients
are only destroyed if they are marked as inactive.