irc: Clarify IRC relay limitations.
Messages will not relay through mod_relay_irc
if the relay channel on the native network is
empty. This can be worked around by guarding
the channel with ChanServ, but this is not
entirely obvious, so this is clarified in the
code and the sample configs.

Some redundant return statements have also been
removed.