mod_sysop: Allow console logging to be toggled on/off.
Sometimes ongoing logging can be annoying when trying
to analyze existing log output. A workaround has been
to type /, which will stop logging while waiting for
the user to finish typing a command, but eventually
this will timeout, and this is not what that is
intended for. This allows the user to explicitly
enable and disable logging to the current console
as desired.