node.c: Add node interrupt capability.
Allow node execution to be "interrupted",
which allows a sysop to manually kick a
node, not entirely, but just from whatever
handler it is currently executing. For
example, if a node is executing a door,
interrupting it will kick it back to
the menu from which it accessed that door.