node.c: Fix leaked nodes at shutdown.
A frequent race condition has been manifesting itself
as the mod_irc_relay test frequently spuriously failing
in the CI because net_irc still has a module reference
and cannot be unloaded. This is due to shutdown_all
removing the node from the node list, but the node
never being freed, thus never relinquishing its reference
on net_irc. This ensures that nodes are always properly
freed rather than aborting erroneously if the node is
no longer in the node list, which it may not be if
the BBS is already shutting down. This should eliminate
most of the spurious test suite failures observed today.

Other changes to reduce test suite flakiness entailed
increasing the allotted time when run under valgrind:

* The time tolerance for test_unit has also been increased,
  due to needing more time on slower machines than was
  previously permitted.
* net_imap TESTLOCK time increased (for slower machines).