mod_webmail: Use stateless sessions, modularize authentication.
mod_webmail now supports stateless sessions, instead of
trying to read PHP session data from localhost. This allows
the frontend web server to be decoupled from the backend
server running net_ws and mod_webmail. For example, the BBS
previously had to coexist with another local web server, say,
Apache HTTP, to run the PHP-based frontend. This can now run
on a different server so PHP no longer needs to be present locally.
Instead of using regular sessions, a cookie is set containing
the server connection info, which the client can resend on each request.
The password is encrypted locally and the encryption key is sent
to the server, which encrypts it and returns it to the client
in a JWE (encrypted JWT). This accomplishes several goals:

* The plaintext password is never stored directly
* The server never stores the plaintext password
* The server is still able to invalidate sessions if needed
  (by changing the JWT encryption key, which would prevent
   the client from recovering its encryption key, unless it
   was stored elsewhere).

This also simplifies server operation since the server no longer
needs to keep track of sessions. This also incidentally fixes
the Remember Me functionality, which never worked properly previously.

A related major security improvement is that this opens the
possibility for authenticating to IMAP servers without the
backend server ever seeing the password (enabled by the previous changes).
This does not apply to plaintext logins, but would apply for challenge
response based schemes. These are not currently supported, but could
be added in the future, such that the client computes the response locally
and only passes the response back through the backend server to the IMAP
server.

Related fixes:

* mod_http: Add HTTP server CLI commands
* net_ws: Fix incorrect argument ordering for route registration
* net.c: Fix not preventing duplicate port registrations
* socket.c: Fail on duplicate port registrations