net_imap: Prohibit flags with unbalanced parentheses.
This is a somewhat tailored fix, but in the wild, flags
have been observed stored in mailboxes with a single (
prepend to the front, e.g. (\Seen. This is obviously
invalid, but must be avoided at all costs, since once
a flag has been written to a mailbox, there is no way to
undo this operation. The flag can be removed from individual
messages, but will generally always be returned in the
PERMANENTFLAGS response for that mailbox for its lifetime.
Due to parsing issues with libetpan in this case, this
can render the mailbox inaccessible. To prevent something
else from shooting a user in the foot this way, we now
prevent the generally unintended writing of malformed flags.