term: Restore original terminal title on exit.
Rather than resetting the terminal title to empty
(or leaving it as is), restore the terminal title
to whatever it was before we overwrote it.

The terminal title is now also more concise
and descriptive.

The title restore is not fully portable between all terminal
emulator types but does work with KiTTY.