socket.c: Don't cast ssize_t to size_t.
bbs_node_readline improperly casted a ssize_t
to size_t, which meant that anytime -1 was
returned, this overflowed to a really large
number, causing memchr to read until it
triggered a segmentation fault.

This is fixed by not casting to size_t here,
and relevant functions have been updated to
return ssize_t instead of int, to make the
return values more semantic.

Fixes #23