net_smtp: Properly reject messages with long lines.
RFC 5322 says that lines in a message cannot exceed
998 characters. We use a sufficiently sized buffer
to accomodate this requirement, but previously
bbs_readline would return the same value if the buffer
was exhausted as if read or poll had failed, and so
the SMTP server would just abruptly disconnect if a
line that was longer than the spec allowed was received.

Now, a different return value is used, allowing the
SMTP server to differentiate buffer exhaustion from
I/O failure, and we return a proper SMTP error code
before disconnecting.