gcc: Use __builtin_unreachable where appropriate.
There are several instances of code, mostly functions
with a switch statement that either have a default or
enumerate all cases explicitly with no default, where
NULL (or some useless value) is returned to satisfy
the function's return value. However, this is dead
code; bbs_assert(0) was frequently used to indicate this.
Instead, use __builtin_unreachable, which is more
semantic, and possibly more performant.