net_telnet: Fix node leak.
For TELNETS, if telnet_handshake fails,
bbs_node_exit was not called so the node
was not automatically cleaned up. The node
would thus stay registered indefinitely
until manually kicked by the sysop.

Node setup for network protocols is also
simplified a bit, and we now keeptrack
of a "lifetime node ID", unique for the lifetime
of the BBS process, to make matching up registered
nodes to previous events in the logs a little
bit easier.