pty.c: Don't attempt to write 0 bytes to node.
For non-ANSI nodes, we call bbs_ansi_strip to
remove all the escape sequences from the buffer
before we write it to the node. However, in the
case that this actually results in an empty
buffer, don't attempt to write it. No nasty
side effects will happen, but a warning will be
thrown otherwise since we shouldn't be attempting
to write 0 bytes.