net_ssh: Fix IP address extraction on failed connect.
When clients fail to negotiate an SSH session (usually
spam connections), the underlying socket file descriptor
was -1 by the time we tried to log it. In fact, we already
have the IP address available to us in a buffer from the
start of the session, so just use that directly for
generating the event, rather than trying to query it
again when it's no longer available to us.