net_msp: Add Message Send Protocol.
This adds support for the Message Send Protocol,
which is a simple protocol that may be used to
receive short messages targeted towards system
users. Both TCP and UDP support is present.
The main intended use case of this is to allow
remote servers to deliver messages to the system
via the IRC and notification subsystems, without
requiring that they maintain persistent TCP
connections to the BBS.