net_irc: Fix truncation of relayed messages.
This improves the IRC message relay process to ensure
that long messages are properly split into multiple
messages if needed to send successfully to IRC.

We also now do more to prevent message truncation
and actively warn about truncation if we detect it.

This address the miscellaneous truncation of various
messages that previously could occur silently without warning.