mod_discord: Add self health check to prevent relay stall.
Occasionally, it seems that the underlying libdiscord library
will stop receiving messages from Discord (sending messages
still works). This results in the relay only working in one
direction, with IRC messages relaying to Discord, but not
vice versa. This has happened several times and is usually
remedied by someone noticing eventually and manually reloading
mod_discord to start everything fresh.

Since we can detect this, until the underlying bug is fixed,
this adds the ability to periodically send test posts to a
test channel and trigger a full module reload if a test post
is not echoed back.