net_smtp: Allow exemptions to STARTTLS requirement.
STARTTLS is typically required to use AUTH commands,
to prevent credentials from leaking on plaintext
connections. However, this may not be a concern on
private local networks, and so the ability to exempt
IPs or CIDR ranges from having to use STARTTLS is added
so that encryption is not needlessly required in these
scenarios.