mod_smtp_delivery_external: Allow relaying mail to/from downstream MTAs.
The ability to accept outgoing mail from other MTAs has been
present for some time. This adds the ability to accept incoming
mail for other MTAs (forwarding it to them) as well as route
an MTA's own outgoing mail through another MTA (smart host).

With this addition, handling of incoming and outgoing mail
is sufficiently flexible to handle most relay scenarios.

* The 'Received' header is now added for such messages as well.

* Fix improper allocation size for delivery status,
and general issues exposed by allowing messages to be
accepted and relayed.

* Eliminate spurious errors for ignored config sections in net_smtp.conf.