mod_sysop: Allow editing commands from history.
Hitherto, it has not been possible to edit commands
retrieved from history (by using the UP or DOWN arrow
keys); it has only been possible to execute them
verbatim, which would appear to be a silly limitation.
This is because we use buffered reading to obtain the
command, and when we are navigating history, we are not
buffered, and the history command is not in the buffer.

To make this a little more user friendly, if the user
presses BACKSPACE, we write the command to the master
end of the console PTY, thus spoofing this as user input,
so that it can be edited naturally.

Some limitations:
* This only works for remote consoles, since an additional
  PTY is not set up for the foreground console, since it
  already has a PTY natively.
* Editing commands from history is triggered exclusively
  by BACKSPACE, i.e. you have to start deleting part of
  a command in history before you can modify it in other ways.
* Users lose the ability to "escape" from history browsing/
  modification by pressing ESC once they have started editing
  a command. Workaround is backspace until it's all gone and
  then hit ENTER; nothing will execute.

Slightly awkward patch, but slightly less user-unfriendly
than it was before.