door_usermgmt: Confirm current password before changing it.
Currently, a user may change his or her password
without confirming the old one first, which is
a security risk. Require the user to confirm the
current password if authenticated, before being
able to proceed to changing it.

This does not apply to unauthenticated changes (resets),
since a verification code by email is used in that case.