mod_smtp_filter_arc: Don't add bare LFs to message.
libopenarc violates the RFCs by producing bare line feeds
in the headers that it generates. To work around this,
we now detect bare line feeds and replace them with CR LF.
This fixes RFC compliance and delivery to MTAs that reject
messages containing bare LFs. This also adds some checking
in other places to detect bare LFs and warn about them.

On top of this, ARC headers were being added to messages
submitted directly for submission (which is unnecessary,
and probably also just wrong). We now only add ARC headers
if the message already had a DKIM signature.