mod_smtp_fetchmail: Add RFC 1985 client.
This complements the existing server support
for the ETRN command by adding a client module
that can log into a remote SMTP server and issue
the ETRN command to flush queues for certain
domains. Because this functionality does not
really intersect with anything in the SMTP server,
it is contained in its own separate module.